+++ test/dy-helper.js
@@ -74,7 +74,7 @@ module.exports.createItems = async ({
       age: Array.isArray(age) === true ? age[i - 1] : age
     };
     // eslint-disable-next-line no-await-in-loop
-    expect(await model.upsert(item)).to.deep.equal({ created: true, item });
+    expect(await model.createOrModify(item)).to.deep.equal({ created: true, item });
     items.push(item);
   }
   return items;
